+++ bindings/nodejs/scripts/neon-build.js
@@ -3,7 +3,6 @@ const { spawnSync } = require('child_process');
 const moveArtifact = require('./move-artifact');
 
 // Passing "--prepack 'npm run build:neon'" causes problems on Windows, so this is a workaround
-
 const npm = process.platform === 'win32' ? 'npm.cmd' : 'npm';
 
 const { status } = spawnSync(npm, ['run', 'build:neon'], {