+++ bindings/nodejs/types/bridge/index.ts
@@ -41,6 +41,7 @@ import type {
     __SignTransactionEssenceMethod__,
     __SubmitAndStoreTransactionMethod__,
     __SyncAccountMethod__,
+    __GetIncomingTransactionDataMethod__,
 } from './account';
 import type {
     __BackupMessage__,
@@ -84,6 +85,7 @@ export type __AccountMethod__ =
     | __GenerateAddressesMethod__
     | __GetBalanceMethod__
     | __GetOutputMethod__
+    | __GetIncomingTransactionDataMethod__
     | __GetFoundryOutputMethod__
     | __GetOutputsWithAdditionalUnlockConditionsMethod__
     | __GetTransactionMethod__