+++ bindings/nodejs/types/index.ts
@@ -4,6 +4,7 @@ export * from './address';
 export * from './bridge';
 export * from './buildOutputData';
 export * from './event';
+export * from './incomingTransactionData';
 export * from './loggerConfig';
 export * from './network';
 export * from './output';