+++ tests/output_preparation.rs
@@ -5,7 +5,7 @@ use std::str::FromStr;
 
 use iota_client::block::address::Address;
 use iota_wallet::{
-    account::{Assets, Features, OutputOptions},
+    account::{Assets, Features, OutputOptions, Unlocks},
     account_manager::AccountManager,
     iota_client::{
         block::output::{NativeToken, NftId, TokenId},
@@ -106,7 +106,7 @@ async fn output_preparation() -> Result<()> {
                 amount: 300000,
                 assets: None,
                 features: Some(Features {
-                    metadata: Some("Hello world".to_string()),
+                    metadata: Some(prefix_hex::encode("Hello world".as_bytes())),
                     tag: None,
                     issuer: None,
                     sender: None,
@@ -123,14 +123,15 @@ async fn output_preparation() -> Result<()> {
     // metadata feature
     assert_eq!(output.features().unwrap().len(), 1);
 
+    // only send 1 with metadata feature
     let output = account
         .prepare_output(
             OutputOptions {
                 recipient_address: recipient_address.clone(),
-                amount: 12000,
+                amount: 1,
                 assets: None,
                 features: Some(Features {
-                    metadata: Some("Hello world".to_string()),
+                    metadata: Some(prefix_hex::encode("Hello world".as_bytes())),
                     tag: None,
                     issuer: None,
                     sender: None,
@@ -141,7 +142,35 @@ async fn output_preparation() -> Result<()> {
             None,
         )
         .await?;
-    assert_eq!(output.amount(), 54600);
+    assert_eq!(output.amount(), 48200);
+    let unlock_conditions = output.unlock_conditions().unwrap();
+    // address + sdr
+    assert_eq!(unlock_conditions.len(), 2);
+    let storage_deposit_return = unlock_conditions.storage_deposit_return().unwrap();
+    // output amount -1
+    assert_eq!(storage_deposit_return.amount(), 48199);
+    // metadata feature
+    assert_eq!(output.features().unwrap().len(), 1);
+
+    let output = account
+        .prepare_output(
+            OutputOptions {
+                recipient_address: recipient_address.clone(),
+                amount: 12000,
+                assets: None,
+                features: Some(Features {
+                    metadata: Some(prefix_hex::encode("Hello world".as_bytes())),
+                    tag: None,
+                    issuer: None,
+                    sender: None,
+                }),
+                unlocks: None,
+                storage_deposit: None,
+            },
+            None,
+        )
+        .await?;
+    assert_eq!(output.amount(), 48200);
     // address and storage deposit unlock condition, because of the metadata feature block, 12000 is not enough for the
     // required storage deposit
     assert_eq!(output.unlock_conditions().unwrap().len(), 2);
@@ -155,7 +184,7 @@ async fn output_preparation() -> Result<()> {
                 amount: 1,
                 assets: None,
                 features: Some(Features {
-                    metadata: Some("Hello world".to_string()),
+                    metadata: Some(prefix_hex::encode("Hello world".as_bytes())),
                     tag: None,
                     issuer: None,
                     sender: None,
@@ -176,8 +205,8 @@ async fn output_preparation() -> Result<()> {
     // metadata feature
     assert_eq!(output.features().unwrap().len(), 1);
 
-    // only works if the nft for this NftId is available in the account
-    if let Ok(output) = account
+    // Error if this NftId is not in the account
+    let error = account
         .prepare_output(
             OutputOptions {
                 recipient_address: recipient_address.clone(),
@@ -186,6 +215,30 @@ async fn output_preparation() -> Result<()> {
                     native_tokens: None,
                     nft_id: Some(NftId::from_str(
                         "0xa068e00a79922eaef241592a7440f131ea7f8ad9e22e580ef139415f273eff30",
+                    )?),
+                }),
+                features: None,
+                unlocks: None,
+                storage_deposit: None,
+            },
+            None,
+        )
+        .await
+        .unwrap_err();
+    match error {
+        iota_wallet::Error::NftNotFoundInUnspentOutputs => {}
+        _ => panic!("should return NftNotFoundInUnspentOutputs error"),
+    }
+
+    if let Ok(output) = account
+        .prepare_output(
+            OutputOptions {
+                recipient_address: recipient_address.clone(),
+                amount: 500000,
+                assets: Some(Assets {
+                    native_tokens: None,
+                    nft_id: Some(NftId::from_str(
+                        "0x0000000000000000000000000000000000000000000000000000000000000000",
                     )?),
                 }),
                 features: None,
@@ -239,6 +292,29 @@ async fn output_preparation() -> Result<()> {
     assert_eq!(features.sender().unwrap().address(), &expected_address);
 
     // error when adding issuer when building basic output
+    let error = account
+        .prepare_output(
+            OutputOptions {
+                recipient_address: recipient_address.clone(),
+                amount: 500000,
+                assets: None,
+                features: Some(Features {
+                    metadata: None,
+                    tag: None,
+                    issuer: Some(issuer_and_sender_address.clone()),
+                    sender: None,
+                }),
+                unlocks: None,
+                storage_deposit: None,
+            },
+            None,
+        )
+        .await
+        .unwrap_err();
+    match error {
+        iota_wallet::Error::MissingParameter(_) => {}
+        _ => panic!("should return MissingParameter error"),
+    }
 
     // issuer and sender address present when building nft output
     let output = account
@@ -267,13 +343,49 @@ async fn output_preparation() -> Result<()> {
     assert_eq!(output.kind(), iota_wallet::iota_client::block::output::NftOutput::KIND);
     assert_eq!(output.amount(), 500000);
     let features = output.features().unwrap();
-    // sender and issuer feature
-    assert_eq!(features.len(), 2);
+    // sender feature
+    assert_eq!(features.len(), 1);
+    let immutable_features = output.immutable_features().unwrap();
+    // issuer feature
+    assert_eq!(immutable_features.len(), 1);
     let issuer_and_sender_address = Address::try_from_bech32(&issuer_and_sender_address)?.1;
-    let issuer_feature = features.issuer().unwrap();
+    let issuer_feature = immutable_features.issuer().unwrap();
     assert_eq!(issuer_feature.address(), &issuer_and_sender_address);
     let sender_feature = features.sender().unwrap();
     assert_eq!(sender_feature.address(), &issuer_and_sender_address);
+
+    // nft with expiration
+    let output = account
+        .prepare_output(
+            OutputOptions {
+                recipient_address: recipient_address.clone(),
+                amount: 500,
+                assets: Some(Assets {
+                    native_tokens: None,
+                    nft_id: Some(NftId::from_str(
+                        "0x0000000000000000000000000000000000000000000000000000000000000000",
+                    )?),
+                }),
+                features: Some(Features {
+                    metadata: None,
+                    tag: None,
+                    issuer: None,
+                    sender: None,
+                }),
+                unlocks: Some(Unlocks {
+                    expiration_unix_time: Some(1),
+                    timelock_unix_time: None,
+                }),
+                // unlocks: None,
+                storage_deposit: None,
+            },
+            None,
+        )
+        .await?;
+    assert_eq!(output.kind(), iota_wallet::iota_client::block::output::NftOutput::KIND);
+    assert_eq!(output.amount(), 53900);
+    // address, sdr, expiration
+    assert_eq!(output.unlock_conditions().unwrap().len(), 3);
 
     std::fs::remove_dir_all("test-storage/output_preparation").unwrap_or(());
     Ok(())